refactor: move hooks into estate.ts
