+++ public/lib/real.ts
@@ -1,4 +1,5 @@
-import { assign, entries, ferr, fmap, fmev, fromEntries } from "./free.ts";
+import { assign, entries, fmap, fromEntries } from "./free.ts";
+import { h, hreg } from "./estate.ts";
 
 export declare namespace JSX {
   export type IntrinsicElements = { [tag: string]: unknown };
@@ -60,112 +61,6 @@ const cj = (cs: Ch[], path: Path): ChE[] =>
  * ```
  */
 
-type Hctx = {
-  hooks: any[];
-  index: number;
-};
-
-const hreg = fmap<string, Hctx>();
-let hctx: Hctx | null = null;
-
-const useHook = <T>(init: () => T) => {
-  if (!hctx) {
-    throw ferr("`useHook` requires a component context");
-  }
-
-  const { hooks, index } = hctx;
-  (index >= hooks.length) && hooks.push(init());
-  hctx.index = index + 1;
-  return hooks[index];
-};
-
-type SetState<T> = (prev: T) => T;
-const isSetState = <T>(x: unknown): x is SetState<T> => typeof x === "function";
-export const useState = <T>(init: T) => {
-  if (!hctx) {
-    throw ferr("`useState` requires a component context");
-  }
-
-  useHook(() => init);
-  const { hooks, index } = hctx;
-
-  const setState = (next: T | SetState<T>) => {
-    hooks[index - 1] = isSetState(next) ? next(hooks[index - 1]) : next;
-    self.dispatchEvent(
-      fmev("message", {
-        data: { type: "render" },
-      }),
-    );
-  };
-
-  return [hooks[index - 1], setState] as const;
-};
-
-type Effect = () => void | (() => void);
-export const useEffect = (effect: Effect, deps?: unknown[]) => {
-  if (!hctx) {
-    throw ferr("`useEffect` requires a component context");
-  }
-
-  const cleanup = useHook(() => null);
-  const prev = useHook(() => null);
-  const { hooks, index } = hctx;
-
-  if (!prev || !deps || deps.some((d, i) => d !== prev[i])) {
-    cleanup?.();
-    hooks[index - 2] = effect() || null;
-    hooks[index - 1] = deps || null;
-  }
-};
-
-export const useMemo = <T>(compute: () => T, deps?: unknown[]) => {
-  if (!hctx) {
-    throw ferr("`useMemo` requires a component context");
-  }
-
-  const memo = useHook(() => null);
-  const prev = useHook(() => null);
-  const { hooks, index } = hctx;
-
-  if (!prev || !deps || deps.some((d, i) => d !== prev[i])) {
-    const next = compute();
-    hooks[index - 2] = next;
-    hooks[index - 1] = deps || null;
-    return next;
-  }
-
-  return memo;
-};
-
-export const useSyncExternalStore = <T>(
-  sub: (cb: () => void) => () => void,
-  get: () => T,
-): T => {
-  if (!hctx) {
-    throw ferr("`useStore` requires a component context");
-  }
-
-  useHook(() => get());
-  useHook(() => null);
-  const { hooks, index } = hctx;
-
-  useEffect(() => {
-    const unsub = sub(() => {
-      hooks[index - 2] = get();
-      self.dispatchEvent(
-        fmev("message", {
-          data: { type: "render" },
-        }),
-      );
-    });
-
-    hooks[index - 1] = unsub;
-    return unsub;
-  }, [sub]);
-
-  return hooks[index - 2];
-};
-
 export const cbs = fmap<string, <E extends Event>(event: E) => void>();
 const hcbs = (props: Props, tid: string): Props =>
   fromEntries(
@@ -189,17 +84,17 @@ export const el = (
   if (tisf(tag)) {
     const hid = String(props?.key ?? path.concat(tag.name).join("."));
 
-    const pctx = hctx;
-    hctx = hreg.get(hid) ?? {
+    const pctx = h.ctx;
+    h.ctx = hreg.get(hid) ?? {
       hooks: [],
       index: 0,
     };
-    hctx.index = 0;
+    h.ctx.index = 0;
 
     const e = ch(tag(props ? hcbs(props, hid) : null, cj(cs, path)), path);
 
-    hreg.set(hid, hctx);
-    hctx = pctx;
+    hreg.set(hid, h.ctx);
+    h.ctx = pctx;
 
     return e;
   }
